[Feature] : Header component design completed
