+++ src/pages/App.jsx
@@ -1,4 +1,4 @@
-import Header from '../component/header/Header';
+import Header from '../components/Header';
 
 const App = () => {
     return <Header />;
